added a size_t cast for linux 32 bit platform build warning

+++ SoarKernel/src/mem.cpp
@@ -51,7 +51,7 @@ void *allocate_memory (agent* thisAgent, size_t size, int usage_code) {
   p = (char *) malloc (size);
   if (p==NULL) {
     char msg[BUFFER_MSG_SIZE];
-    SNPRINTF(msg, BUFFER_MSG_SIZE, "\nmem.c: Error:  Tried but failed to allocate %lu bytes of memory.\n", size);
+    SNPRINTF(msg, BUFFER_MSG_SIZE, "\nmem.c: Error:  Tried but failed to allocate %lu bytes of memory.\n", static_cast<unsigned long>(size));
 	msg[BUFFER_MSG_SIZE - 1] = 0; /* ensure null termination */
     abort_with_fatal_error (thisAgent, msg);
   }
